freertos: rename esp32 plat to freertos
