+++ lib/core-net/adopt.c
@@ -400,7 +400,7 @@ lws_create_adopt_udp(struct lws_vhost *vhost, int port, int flags,
 	lws_snprintf(buf, sizeof(buf), "%u", port);
 	n = getaddrinfo(NULL, buf, &h, &r);
 	if (n) {
-#ifndef LWS_WITH_ESP32
+#if !defined(LWS_PLAT_FREERTOS)
 		lwsl_info("%s: getaddrinfo error: %s\n", __func__, gai_strerror(n));
 #else
         lwsl_info("%s: getaddrinfo error: %s\n", __func__, strerror(n));
