+++ lib/core/libwebsockets.c
@@ -69,7 +69,7 @@ lws_hex_to_byte_array(const char *h, uint8_t *dest, int max)
 
 #if !defined(LWS_PLAT_OPTEE)
 
-#if !defined(LWS_AMAZON_RTOS)
+#if defined(LWS_WITH_FILE_OPS)
 int lws_open(const char *__file, int __oflag, ...)
 {
 	va_list ap;
