+++ lib/plat/freertos/freertos-helpers.c
