+++ lib/plat/freertos/freertos-init.c
