+++ lib/plat/freertos/freertos-service.c
