+++ lib/roles/ws/ops-ws.c
@@ -903,7 +903,7 @@ lws_server_init_wsi_for_ws(struct lws *wsi)
 	wsi->ws->rx_ubuf_alloc = n;
 	lwsl_debug("Allocating RX buffer %d\n", n);
 
-#if !defined(LWS_WITH_ESP32)
+#if !defined(LWS_PLAT_FREERTOS)
 	if (!wsi->h2_stream_carries_ws)
 		if (setsockopt(wsi->desc.sockfd, SOL_SOCKET, SO_SNDBUF,
 		       (const char *)&n, sizeof n)) {
