+++ lib/tls/mbedtls/mbedtls-ssl.c
@@ -56,7 +56,7 @@ lws_ssl_capable_read(struct lws *wsi, unsigned char *buf, int len)
 
 	errno = 0;
 	n = SSL_read(wsi->tls.ssl, buf, len);
-#if defined(LWS_WITH_ESP32)
+#if defined(LWS_PLAT_FREERTOS)
 	if (!n && errno == LWS_ENOTCONN) {
 		lwsl_debug("%p: SSL_read ENOTCONN\n", wsi);
 		return LWS_SSL_CAPABLE_ERROR;
