+++ lib/tls/private-lib-tls.h
@@ -52,7 +52,7 @@
   #define OPENSSL_NO_TLSEXT
  #endif /* not USE_OLD_CYASSL */
 #else /* WOLFSSL */
- #if defined(LWS_WITH_ESP32)
+ #if defined(LWS_PLAT_FREERTOS)
   #define OPENSSL_NO_TLSEXT
   #if !defined(LWS_AMAZON_RTOS)
    /* AMAZON RTOS has its own setting via MTK_MBEDTLS_CONFIG_FILE */
